server/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.0
txtai>=5.0.0
python-multipart>=0.0.5
typing-extensions>=4.0.0
